pyproject.toml
[tool.poetry]
name = "aws-event-driven-pipeline"
version = "0.1.0"
description = "Event-driven ingestion and analytics pipeline"
authors = ["simoda"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.180.0"
constructs = "^10.0.0"
boto3 = "^1.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
